backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
supabase==2.11.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
sentence-transformers>=3.0.0
httpx>=0.27.0
